requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
transformers
torch
accelerate
python-jose
passlib
python-multipart
loguru
python-dotenv